+++ docs/dev-demos/src/api/glyph.ts
@@ -196,12 +196,12 @@ export const callback = (chartInstance: any) => {
   increaseButton.addEventListener('click', () => {
     const signal = chartInstance.getSignalById('filterSignal');
     signal.value(signal.output() + 1);
-    chartInstance.runAsync();
+    chartInstance.run();
   });
   decreaseButton.addEventListener('click', () => {
     const signal = chartInstance.getSignalById('filterSignal').value();
     signal.value(signal.output() - 1);
-    chartInstance.runAsync();
+    chartInstance.run();
   });
   disappearButton.addEventListener('click', () => {
     chartInstance.getSignalById('animationState').value('disappear');