+++ docs/dev-demos/src/specs/roam-scale-bar.ts
@@ -262,7 +262,7 @@ export const callback = (view: IView) => {
       { name: 'L', value: Math.floor(100 * Math.random()) },
       { name: 'M', value: Math.floor(100 * Math.random()) }
     ]);
-    view.runAsync();
+    view.run();
   });
 };